# backend/app/static/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/ml/seasons": {
            "post": {
                "summary": "Seasons",
                "description": "Requirements in json {\n    seg_class_code: str,\n    flt_num: int,\n}\nNot require, but can be used in json {\n    date_start: YYYY-MM-DD = -1-1,\n    date_finish: YYYY-MM-DD = -12-31,\n    period: int = 365,\n    fourier: int | None = None,\n}",
                "operationId": "seasons_ml_seasons_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeasonsReport"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Seasons Ml Seasons Post",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/SeasonsResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ml/dynamic": {
            "post": {
                "summary": "Dynamic",
                "description": "Requirements in json {\n    seg_class_code: str,\n    flt_num: int,\n    dep_date: YYYY-MM-DD,\n}\nNot require, but can be used in json {\n    period_start: str | None = None,\n    period_end: str | None = None,\n    fourier: int | None = None,\n}",
                "operationId": "dynamic_ml_dynamic_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DynamicReport"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Dynamic Ml Dynamic Post",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/DynamicResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ml/predict": {
            "post": {
                "summary": "Predict",
                "description": "Requirements in json {\n    seg_class_code: str,\n    flt_num: int,\n    dep_date: YYYY-MM-DD,\n}\nNot require, but can be used in json {\n    dtd_start: int = -1,\n    dtd_end: int = 30,\n}",
                "operationId": "predict_ml_predict_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PredictReport"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Predict Ml Predict Post",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/DynamicResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "DynamicReport": {
                "title": "DynamicReport",
                "required": [
                    "seg_class_code",
                    "flt_num",
                    "dep_date"
                ],
                "type": "object",
                "properties": {
                    "seg_class_code": {
                        "title": "Seg Class Code",
                        "type": "string"
                    },
                    "flt_num": {
                        "title": "Flt Num",
                        "type": "integer"
                    },
                    "dep_date": {
                        "title": "Dep Date",
                        "type": "string"
                    },
                    "period_start": {
                        "title": "Period Start",
                        "type": "string"
                    },
                    "period_end": {
                        "title": "Period End",
                        "type": "string"
                    },
                    "fourier": {
                        "title": "Fourier",
                        "type": "integer"
                    }
                }
            },
            "DynamicResponse": {
                "title": "DynamicResponse",
                "required": [
                    "flight_dynamic"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Status"
                            }
                        ],
                        "default": "ok"
                    },
                    "message": {
                        "title": "Message",
                        "type": "string"
                    },
                    "flight_dynamic": {
                        "title": "Flight Dynamic",
                        "type": "array",
                        "items": {}
                    },
                    "fourier_dynamic": {
                        "title": "Fourier Dynamic",
                        "type": "array",
                        "items": {}
                    }
                }
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "type": "object",
                "properties": {
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Status"
                            }
                        ],
                        "default": "error"
                    },
                    "message": {
                        "title": "Message",
                        "type": "string"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "PredictReport": {
                "title": "PredictReport",
                "required": [
                    "seg_class_code",
                    "flt_num",
                    "dep_date"
                ],
                "type": "object",
                "properties": {
                    "seg_class_code": {
                        "title": "Seg Class Code",
                        "type": "string"
                    },
                    "flt_num": {
                        "title": "Flt Num",
                        "type": "integer"
                    },
                    "dep_date": {
                        "title": "Dep Date",
                        "type": "string"
                    },
                    "dtd_start": {
                        "title": "Dtd Start",
                        "type": "integer",
                        "default": -1
                    },
                    "dtd_end": {
                        "title": "Dtd End",
                        "type": "integer",
                        "default": 30
                    }
                }
            },
...
            "SeasonsReport": {
                "title": "SeasonsReport",
                "required": [
                    "seg_class_code",
                    "flt_num"
                ],
                "type": "object",
                "properties": {
                    "seg_class_code": {
                        "title": "Seg Class Code",
                        "type": "string"
                    },
                    "flt_num": {
                        "title": "Flt Num",
                        "type": "integer"
                    },
                    "date_start": {
                        "title": "Date Start",
                        "type": "string",
                        "default": "2018-01-01"
                    },
                    "date_finish": {
                        "title": "Date Finish",
                        "type": "string",
                        "default": "2018-12-31"
                    },
                    "period": {
                        "title": "Period",
                        "type": "integer",
                        "default": 365
                    },
                    "fourier": {
                        "title": "Fourier",
                        "type": "integer"
                    }
                }
            },
            "SeasonsResponse": {
                "title": "SeasonsResponse",
                "required": [
                    "seasons"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Status"
                            }
                        ],
                        "default": "ok"
                    },
                    "message": {
                        "title": "Message",
                        "type": "string"
                    },
                    "seasons": {
                        "title": "Seasons",
                        "type": "array",
                        "items": {}
                    },
                    "fourier_seasons": {
                        "title": "Fourier Seasons",
                        "type": "array",
                        "items": {}
                    }
                }
            },
            "Status": {
                "title": "Status",
                "enum": [
                    "ok",
                    "error",
                    "in_progress"
                ],
                "description": "An enumeration."
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}